requirements.txt
tensorflow>=2.20.0
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0
matplotlib>=3.7.0
scikit-learn>=1.3.0
flask>=2.3.0
flask-cors>=4.0.0
werkzeug>=2.3.0
gunicorn
